some restructure.. again
